+++ src/Resource.php
@@ -37,26 +37,11 @@ class Resource
     public static $viewsDir = 'views';
 
     /**
-     * Callback to handle exceptions
-     * Should not throw exception
-     *
-     * @var callback
-     */
-    public static $onError = 'error_log';
-
-    /**
      * Should the layout be used in render.
      *
      * @var boolean
      */
     public static $layout = true;
-
-    /**
-     * List of allowed HTTP methods.
-     *
-     * @var array
-     */
-    public static $allowedRequestMethods = array('GET', 'POST', 'PUT', 'DELETE', 'HEAD');
 
     /**
      * Last modified timestamp
@@ -73,104 +58,6 @@ class Resource
     public $error;
 
     /**
-     * Route to a matching resource, calling the appropriate
-     * HTTP method and render() the response.
-     *
-     * @param array of \Http\Resource
-     * @param callback \Http\Resource factory function, default to `new`
-     * @return void
-     */
-    public static function handle(array $resources, $factory = null)
-    {
-        try {
-            $factory = $factory ?: function ($className) { return new $className(); };
-            $uri = parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH);
-
-            // safeguard, parse_url should work
-            if ($uri === false) {
-                throw new NotFound();
-            }
-
-            // begin opinionated here ...
-            if (preg_match('|.+/$|', $uri)) {
-                throw new MovedPermanently( rtrim($uri, '/') . (empty($_SERVER['QUERY_STRING']) ? "" : "?{$_SERVER['QUERY_STRING']}") );
-            }
-
-            // method override
-            if (isset($_GET['_method'])) {
-                $_SERVER['REQUEST_METHOD'] = strtoupper($_GET['_method']);
-            }
-
-            if (!in_array($_SERVER['REQUEST_METHOD'], static::$allowedRequestMethods)) {
-                throw new MethodNotAllowed();
-            }
-
-            // lookup $resources and call appropriate method
-            foreach ($resources as $className) {
-                $props = $className::match($uri);
-                if ($props !== false) {
-                    $resource = call_user_func($factory, $className);
-                    $resource->dispatch($_SERVER['REQUEST_METHOD'], $props);
-                    $resource->render();
-                    break;
-                }
-            }
-
-            if (empty($resource)) {
-                throw new NotFound();
-            }
-
-        } catch (NotModified $e) {
-            // don't output anything
-        } catch (Redirection $e) {
-            header("Location: $e->location", true, $e->code);
-            echo $e->getMessage();
-        } catch (Exception $e) {
-            // fallback to default view rendering
-        } catch (\Exception $e) {
-            $fn = isset($resource) ? $resource::$onError : static::$onError;
-            call_user_func($fn, $e);
-            $e = new InternalServerError($e->getMessage(), $e);
-        }
-        /* finally */ if (isset($e)) {
-            header("{$_SERVER['SERVER_PROTOCOL']} $e->code $e->reason");
-            if ($e instanceof \Http\Error) {
-                // not rendering redirects
-                $error = new static();
-                $error->error = $e;
-                $error->renderTwoStepView($e);
-            }
-        }
-    }
-
-    /**
-     * Initialize resource with $properties and call appropiate $method.
-     *
-     * @param string $method
-     * @param array $properties
-     */
-    public function dispatch($method, array $properties)
-    {
-        if (!method_exists($this, $method)) {
-            throw new MethodNotAllowed();
-        }
-        // assign properties to resource and initialize
-        foreach ($properties as $key => $value) {
-            $this->$key = $value;
-        }
-        $this->init();
-        $this->$method();
-        // caching headers
-        if ($this->lastModified) {
-            $lastModified = gmdate('r', $this->lastModified);
-            header("Last-Modified: $lastModified");
-            if (isset($_SERVER['HTTP_IF_MODIFIED_SINCE']) && $_SERVER['HTTP_IF_MODIFIED_SINCE'] == $lastModified) {
-                throw new NotModified(null);
-            }
-        }
-    }
-
-    /**
      * Test if the resource path match.
      *
      * @param string Request-URI to match against
@@ -178,13 +65,13 @@ class Resource
      */
     public static function match($uri)
     {
-        $replacements = array(
+        $replacements = [
             '/\(/' => '(',
             '/\)/' => ')?',
             '/\./' => '\.',
             '/:(\w+)/' => '(?P<$1>[^./]+)',
             '/\*/' => '(?P<rest>.*?)',
-        );
+        ];
         $path = preg_replace('/^\//', '/?', static::$path);
         $route = static::$base . $path;
         foreach ($replacements as $pattern => $replacement) {
@@ -236,11 +123,25 @@ class Resource
     }
 
     /**
-     * Called right after properties are assigned.
+     * Called right before method call.
      */
     public function init()
     {
         ;
+    }
+
+    /**
+     * Called right after method call.
+     */
+    public function finalize()
+    {
+        if ($this->lastModified) {
+            $lastModified = gmdate('r', $this->lastModified);
+            header("Last-Modified: $lastModified");
+            if (isset($_SERVER['HTTP_IF_MODIFIED_SINCE']) && $_SERVER['HTTP_IF_MODIFIED_SINCE'] == $lastModified) {
+                throw new NotModified(null);
+            }
+        }
     }
 
     /**
@@ -256,7 +157,7 @@ class Resource
      */
     public function get()
     {
-        throw new \Http\MethodNotAllowed();
+        throw new MethodNotAllowed();
     }
 
     /**
@@ -295,7 +196,7 @@ class Resource
                 $name = dirname($name);
                 $file = static::$viewsDir . "/$name/layout.php";
                 if (file_exists($file)) {
-                    $content = $this->partial($file, array('content' => $content));
+                    $content = $this->partial($file, ['content' => $content]);
                     break;
                 }
             } while ($name != '.');